+++ WebContent/js/components/JobInstance.js
@@ -14,7 +14,7 @@ import { Map, List } from 'immutable';
 import { connect } from 'react-redux';
 import LabelIcon from '@material-ui/icons/Label';
 import { ContextMenu, MenuItem, ContextMenuTrigger } from 'react-contextmenu';
-import { fetchJobFiles, toggleJob, purgeJob } from '../actions/jobNodes';
+import { fetchJobFiles, toggleJob, cancelJob, purgeJob } from '../actions/jobNodes';
 import { getJCL, getFileLabel, changeSelectedContent, fetchConcatenatedJobFiles } from '../actions/content';
 import JobFile from './JobFile';
 import JobStep from './JobStep';
@@ -100,6 +100,11 @@ class JobInstance extends React.Component {
         dispatch(purgeJob(job.get('jobName'), job.get('jobId')));
     }
 
+    handleCancel(job) {
+        const { dispatch } = this.props;
+        dispatch(cancelJob(job.get('jobName'), job.get('jobId')));
+    }
+
     handleGetJCL(job) {
         const { dispatch } = this.props;
         const fileLabel = getFileLabel(job.get('jobId'), 'JCL');
@@ -155,17 +160,26 @@ class JobInstance extends React.Component {
 
     renderJobInstanceMenu() {
         const { job } = this.props;
+        const menuItems = [
+            <MenuItem key="open" onClick={() => { this.handleOpenAllFiles(job); }}>
+                    Open
+            </MenuItem>,
+            <MenuItem key="purge" onClick={() => { this.handlePurge(job); }}>
+            Purge Job
+            </MenuItem>,
+            <MenuItem key="getJCL" onClick={() => { this.handleGetJCL(job); }}>
+                Get JCL (SJ)
+            </MenuItem>,
+        ];
+        if (job.get('status').toLowerCase() === 'active') {
+            menuItems.splice(1, 0,
+                <MenuItem key="cancel" onClick={() => { this.handleCancel(job); }}>
+                    Cancel Job
+                </MenuItem>);
+        }
         return (
             <ContextMenu id={job.get('label')}>
-                <MenuItem onClick={() => { this.handleOpenAllFiles(job); }}>
-                    Open
-                </MenuItem>
-                <MenuItem onClick={() => { this.handlePurge(job); }}>
-                    Purge Job
-                </MenuItem>
-                <MenuItem onClick={() => { this.handleGetJCL(job); }}>
-                    Get JCL (SJ)
-                </MenuItem>
+                {menuItems}
             </ContextMenu>
         );
     }